银行放款记录录入

+++ src/main/java/com/yunche/loan/domain/vo/BankLendRecordVO.java
@@ -46,6 +46,6 @@ public class BankLendRecordVO {
     /**
      * 放款金额
      */
-    private double lendAmount;
+    private BigDecimal lendAmount;
 
 }
